Consolidated all S3 payload POJO into class S3Payload
